refactor: extract reusable slide confirm widget
